refactor: copyToTeamDialog
