+++ apps/journeys-admin/src/components/JourneyList/JourneyCard/JourneyCardMenu/DuplicateJourneyMenuItem/DuplicateJourneyMenuItem.tsx
@@ -10,9 +10,9 @@ import { setBeaconPageViewed } from '@core/journeys/ui/setBeaconPageViewed'
 import { useJourneyDuplicateMutation } from '@core/journeys/ui/useJourneyDuplicateMutation'
 import CopyLeftIcon from '@core/shared/ui/icons/CopyLeft'
 
+import { CopyToTeamDialog } from '@core/journeys/ui/CopyToTeamDialog'
 import { JourneyDuplicate } from '../../../../../../__generated__/JourneyDuplicate'
 import { MenuItem } from '../../../../MenuItem'
-import { CopyToTeamDialog } from '../../../../Team/CopyToTeamDialog'
 
 interface DuplicateJourneyMenuItemProps {
   id?: string
