+++ apps/journeys-admin/src/components/Team/CopyToTeamMenuItem/CopyToTeamMenuItem.tsx
@@ -7,8 +7,8 @@ import { setBeaconPageViewed } from '@core/journeys/ui/setBeaconPageViewed'
 import { useJourneyDuplicateMutation } from '@core/journeys/ui/useJourneyDuplicateMutation'
 import CopyToIcon from '@core/shared/ui/icons/CopyTo'
 
+import { CopyToTeamDialog } from '@core/journeys/ui/CopyToTeamDialog'
 import { MenuItem } from '../../MenuItem'
-import { CopyToTeamDialog } from '../CopyToTeamDialog'
 
 interface DuplicateJourneyMenuItemProps {
   id?: string
