+++ libs/journeys/ui/src/components/CopyToTeamDialog/CopyToTeamDialog.stories.tsx
@@ -2,12 +2,12 @@ import { MockedResponse } from '@apollo/client/testing'
 import { Meta, StoryObj } from '@storybook/react'
 import { SnackbarProvider } from 'notistack'
 
+import { journeysAdminConfig } from '@core/shared/ui/storybook'
 import {
   GET_LAST_ACTIVE_TEAM_ID_AND_TEAMS,
   TeamProvider
-} from '@core/journeys/ui/TeamProvider'
-import { GetLastActiveTeamIdAndTeams } from '@core/journeys/ui/TeamProvider/__generated__/GetLastActiveTeamIdAndTeams'
-import { journeysAdminConfig } from '@core/shared/ui/storybook'
+} from '../TeamProvider'
+import { GetLastActiveTeamIdAndTeams } from '../TeamProvider/__generated__/GetLastActiveTeamIdAndTeams'
 
 import { CopyToTeamDialog } from './CopyToTeamDialog'
